+++ includes/function.php
@@ -1,8 +1,8 @@
 <?php
-function TttPrice($price) {
+function ttcPrice($price) {
     return $price * 0.8;
 }
-function DisplayProduct($key,Array $product){
+function displayProduct($key,Array $product){
             echo
             "<tr>
             <td>" . $key . "</td>
@@ -13,8 +13,8 @@ function DisplayProduct($key,Array $product){
             } else {
                 echo'style="color:blue"';
             }
-            echo ">" . $product['price'] . "</td>
-            <td>" . TttPrice($product['price']) . "</td>
+            echo ">" . $product['price'] . " €</td>
+            <td>" . ttcPrice($product['price']) . " €</td>
             <td>" . $product['desc'] . "</td>
             </tr>";
 }
