+++ includes/variables.php
@@ -8,7 +8,7 @@
          'price' => 14,
           'desc' =>  "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam." 
 ],
-        ['name'  => "Bonnet en laine et cachemire",
+        ['name'  => "Bonnet en cachemire",
          'price' => 20,
           'desc' =>  "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam." 
 ],
